@@ -9,6 +9,7 @@ use Rareloop\Lumberjack\Providers\ServiceProvider;
 use HelloNico\Twig\DumpExtension;
 use Ajgl\Twig\Extension\BreakpointExtension;
 use Djboris88\Twig\Extension\CommentedIncludeExtension;
+use Timber\Twig_Function;
 
 /**
  * Class TwigExtensionsServiceProvider
@@ -20,20 +21,15 @@ class TwigExtensionsServiceProvider extends ServiceProvider
      * Register all extensions listed into the config file
      * @param Config $config
      */
-    public function boot(Config $config)
+    public function boot(Config $config): void
     {
         add_filter('timber/twig', function ($twig) use ($config) {
             $functionsToRegister = $config->get('twig.allowed_functions');
-            //$twig->addExtension(ShareExtension::getInstance());
 
-            $phpFunctions = array_merge([
-                "mix",
-                "mix_any"
-            ], $functionsToRegister ?? []);
-
-            $twig->addExtension(new \Umpirsky\Twig\Extension\PhpFunctionExtension($phpFunctions));
-            $twig->addExtension(new \DPolac\TwigLambda\LambdaExtension());
             $twig->addExtension(new \Aaronadal\TwigListLoop\Twig\TwigExtension());
+            foreach ($functionsToRegister ?? [] as $function) {
+                $twig->addFunction(new Twig_Function($function, $function));
+            }
 
             if (defined("WP_DEBUG") && WP_DEBUG) {
                 $twig->addExtension(new CommentedIncludeExtension());
@@ -49,9 +45,7 @@ class TwigExtensionsServiceProvider extends ServiceProvider
              * Adding a second filter to cover the `Timber::render()` case, when the
              * template is not loaded through the `include` tag inside a twig file
              */
-            add_filter( 'timber/output', function( $output, $data, $file ) {
-                return "\n<!-- Begin output of '" . $file . "' -->\n" . $output . "\n<!-- / End output of '" . $file . "' -->\n";
-            }, 10, 3 );
+            add_filter('timber/output', fn($output, $data, $file) => "\n<!-- Begin output of '" . $file . "' -->\n" . $output . "\n<!-- / End output of '" . $file . "' -->\n", 10, 3);
         }
 
         $extensionsToRegister = $config->get('twig.extensions');
