@@ -3,25 +3,25 @@
 namespace Adeliom\WP\Extensions\Utils\Types;
 
 /**
- * Class Post_Slug
+ * Class PostSlug
  *
  * Allows you to register callback functions to customize post slugs.
  */
-class Post_Slug
+class PostSlug
 {
     /**
      * Post type callbacks.
      *
      * @var array An associative array of post types and their callbacks.
      */
-    private $post_types = [];
+    private array $post_types = [];
 
     /**
      * Inits hooks.
      */
-    public function init()
+    public function init(): void
     {
-        add_filter('wp_insert_post_data', [$this, 'customize_slug'], 11, 2);
+        add_filter('wp_insert_post_data', fn(array $data, array $postarr): array => $this->customizeSlug($data, $postarr), 11, 2);
     }
 
     /**
@@ -29,7 +29,7 @@ class Post_Slug
      *
      * @param array $post_types An associative array of post types and their callbacks.
      */
-    public function register($post_types = [])
+    public function register(array $post_types = []): void
     {
         foreach ($post_types as $post_type => $callback) {
             $this->post_types[$post_type] = [
@@ -45,9 +45,9 @@ class Post_Slug
      * @param array $data An array of slashed post data.
      * @param array $postarr An array of sanitized, but otherwise unmodified post data.
      *
-     * @return array
+     * @return mixed[]
      */
-    public function customize_slug($data, $postarr)
+    public function customizeSlug(array $data, array $postarr): array
     {
         $bailout_states = ['auto-draft', 'trash'];
         $post_status    = $postarr['post_status'];
@@ -60,7 +60,8 @@ class Post_Slug
         $post_type = $postarr['post_type'];
 
         // Bailout if no callback could be found.
-        if (!in_array($post_type, array_keys($this->post_types), true)
+        if (
+            !array_key_exists($post_type, $this->post_types)
             || !is_callable($this->post_types[$post_type]['callback'])
         ) {
             return $data;
