@@ -2,7 +2,6 @@
 
 namespace Adeliom\WP\Extensions\Providers;
 
-use Dugajean\WpHookAnnotations\HookRegistry;
 use Rareloop\Lumberjack\Config;
 use Rareloop\Lumberjack\Providers\ServiceProvider;
 
@@ -13,26 +12,31 @@ use Rareloop\Lumberjack\Providers\ServiceProvider;
 class HookServiceProvider extends ServiceProvider
 {
     /**
-     * Bind HookRegistry into the container
-     */
-    public function register()
-    {
-        $hookRegisty = new HookRegistry();
-        $this->app->bind("hooks", $hookRegisty);
-        $this->app->bind(HookRegistry::class, $hookRegisty);
-    }
-
-    /**
      * Register all hooks listed into the config file
      * @param Config $config
+     * @throws \ReflectionException
      */
-    public function boot(Config $config)
+    public function boot(Config $config): void
     {
         $hooksToRegister = $config->get('hooks.register');
-        if (is_array($hooksToRegister)) {
-            foreach ($hooksToRegister as $hook) {
-                $this->app->get(HookRegistry::class)->bootstrap($hook);
+        $ns = [];
+        foreach ($hooksToRegister as $hook) {
+            $class = new \ReflectionClass($hook);
+            $hookNs = $class->getNamespaceName();
+            if (!in_array($hookNs, $ns, true)) {
+                $ns[] = $hookNs;
             }
         }
+
+        add_filter('wp_hook_attributes_registered_namespaces', function () use ($ns) {
+            return $ns;
+        });
+
+        add_filter('wp_hook_attributes_registered_classes', function (array $registered_classes) use ($hooksToRegister): array {
+            return array_merge(
+                $registered_classes,
+                $hooksToRegister
+            );
+        });
     }
 }
